Safari-611.1.10/LayoutTests/css3/flexbox/flexbox-ignore-firstLetter.html

<!DOCTYPE html>
<html>
<link href="resources/flexbox.css" rel="stylesheet">
<style>
    div::first-letter { line-height: 100px; }
    p { line-height: 20px; }
</style>

<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/check-layout-th.js"></script>
<body onload="checkLayout('.flexbox, .inline-flexbox')">
<div id=log></div>
<p>This test flex item should ignore flex container's firstLetter pseudo element.</p>

<div class="flexbox">
    <p data-expected-height=20>The first item.</p>
    <p data-expected-height=20>The second item.</p>
</div>
<div class="inline-flexbox">
    <p data-expected-height=20>The first item.</p>
    <p data-expected-height=20>The second item.</p>
</div>
</body>
</html>
